+++ database/factories/CompanyFactory.php
@@ -2,6 +2,8 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
+namespace Database\Factories;
+
 use App\{Company, Turn};
 use Faker\Generator as Faker;
 use Illuminate\Support\Str;
