+++ database/factories/SectionFactory.php
@@ -2,6 +2,8 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
+namespace Database\Factories;
+
 use App\Section;
 use App\Source;
 use Faker\Generator as Faker;
