+++ database/factories/TurnFactory.php
@@ -2,6 +2,8 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
+namespace Database\Factories;
+
 use App\Turn;
 use Faker\Generator as Faker;
