+++ database/factories/UserFactory.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Factories;
+
 use Faker\Generator as Faker;
 use Illuminate\Support\Str;
