+++ database/seeders/SectionSeeder.php
@@ -1,12 +1,14 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Section;
 use Illuminate\Database\Seeder;
 
 class SectionSeeder extends Seeder
 {
     /**
-     * Run the database seeds.
+     * Run the database seeders.
      *
      * @return void
      */
