+++ database/seeders/SourceSeeder.php
@@ -1,12 +1,14 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Source;
 use Illuminate\Database\Seeder;
 
 class SourceSeeder extends Seeder
 {
     /**
-     * Run the database seeds.
+     * Run the database seeders.
      *
      * @return void
      */
