+++ public/index.php
@@ -9,6 +9,9 @@
 
 define('LARAVEL_START', microtime(true));
 
+if (file_exists(__DIR__.'/../storage/framework/maintenance.php')) {
+    require __DIR__.'/../storage/framework/maintenance.php';
+}
 /*
 |--------------------------------------------------------------------------
 | Register The Auto Loader
